New NFSD Ships: Zweihander and Stallion
Two new NFSD ships to help with the lack of variety in their shipyard, done... about as well as I can on my own. Please kill me with several hundred rocks to the head if anything here is incorrect or out of place.

If everything works, you should be able to test the ships just by buying them at the NFSD shipyard computer.